Initial commit of unmodified vbcc.
This is the UNMODIFIED source distribution of Volker Barthelmann's
vbcc compiler. This includes a Z-machine backend originally by
David Given (2001) and updated by Volker Barthelmann (2006).

For now, I'll follow Given's lead in distributing my modifications
to vbcc as a big patch, even though you could certainly argue that
any modifications made via git are already a series of patches.
I have contacted Dr. Barthelmann to ask for clarification and/or
permission to distribute a modified version in this case.