Optimize puts("foo") into the Z-machine's @print opcode.
Now we begin hacking vbcc in earnest. I'm inserting this optimization
right into the indep "ic.c", just because there doesn't seem to be
any elegant place to put it. (I don't think HAS_LIBCALLS is quite
what I'm looking for, unfortunately.)

After this change, "advent.c" will successfully compile AND assemble,
producing "advent.z8". You can even run it in an interpreter (I'm
using fizmo as my interpreter for now), and see it produce a sort of
lopsided version of Adventure --- a lot of things work, but there are
a LOT of bugs and glitches that don't appear in the gcc-compiled
version. My next task will be to track down and eradicate the
compiler bugs that I presume are responsible.

New changes to vbcc-patch.diff in this revision are:

    - In "ic.c", look for CALL-expressions like puts("hello")
      and replace them with calls to made-up inline_asm functions
      that will expand into the right thing, namely, a single
      @print opcode. This drastically cuts down on static data
      size, which means we can assemble larger C programs.