Fix a bunch of long/int format-string mismatches.
These mismatches showed up only on 32-bit systems, where they caused the
compiler to produce completely bogus Inform code: usually in the form of
variable names showing up as "(null)" because there was unexpectedly an
extra zero word on the stack from the previous "zlong" or "zmax" value.

On a 32-bit Ubuntu system:
  zmax is zllong is long long
  zlong is int(!)
  zint is short(!)
  zshort is short
  zchar is signed char

On a 64-bit Ubuntu system:
  zmax is zllong is long(!)
  zlong is int(!)
  zint is short(!)
  zchar is signed char

So basically vbcc likes to pretend that it's running on a 16-bit system
with 32-bit longs, but it will avoid using "long long" if "long" is
a 64-bit type already. What a mess.

In general, I am fixing this by using the plain old C types "short" and
"int". Since the Z-machine itself is only a 16-bit machine, there's no
reason ever to use "long" or "long long" in the backend.