Fix a bug in the generation of the __init_vars function.
The routine cleanup_cg() emits the __init_vars_<modulename>() function to
initialize all the static data containing pointers, since that can't be
done statically. It's using @storew instructions to store the data, but
it forgot to use word-offsets instead of byte-offsets: i.e., there was
a missing ">> 1".

After this change, the initial "Welcome to Colossal Cave!!" help-text
displays correctly. The turn-counter still displays as garbage when
you quit, though, and "xyzzy" from the Well House will take you to Y2
instead of to the debris room. That's a weird bug.