Fix my off-by-one error in dealing with MAX_SPEC.
Knuth apparently leaves remarks[0] empty, and begins the remarks
table at remarks[1]. I had missed that, and started my remarks at 0,
which meant that the logic for when to print a remark was wrong,
because I had MAX_SPEC+0 meaning "print remark 0" at the same time
that it meant "go to R_TROLL".

This bug was exposed whenever the game should have printed remark 0,
which is the "you can't fit through a two-inch slit" message.