Add a copy of Parchment to the repository.
Typing "make parchment" will now copy advent.z8 into the parchment/
directory, so that you can run it by pointing your Web browser to
parchment/index.html (even in offline mode).

I acquired this copy of Parchment by packaging a dummy Inform 7
project for "release", and then switching out the project-specific
bits. Notice that index.html doesn't mention the Treaty of Babel
IFID for advent.z8; that's partly because it's a work in progress,
and partly because I'm not sure how to choose one. (I suspect that
it's just the md5sum of advent.z8, but maybe the Inform assembler
is inserting an IFID for me already.)

I've also added the commands SAVE and RESTO[re], which will eventually
do what it says on the tin, but for now are slightly broken. I need
to understand what the Z-machine means by "save" and "restore" first.
SAVE and RESTORE are not available in the gcc-compilable version,
because that's not a very interesting problem to me.