First commit of quuxplusone.github.com/Advent/
If I'm doing this right, this "gh-pages" branch should allow you
to play Adventure right in your browser from a GitHub URL, rather
than needing to download the .z8 file or the whole repository.

This branch will probably end up out-of-date as I make changes
to the parchment/ subdirectory of the "master" branch and forget
to merge them here manually, but if this works, it'll be pretty
cool anyway.