Fix a major backend bug with "stackparamadjust".
To compile a call to a variadic function, the middle-end generates
a series of PUSHes followed by a CALL. It doesn't generate any POPs;
it leaves that up to the backend.

We do need to generate POPs to get the stack pointer (that's "xp",
not "sp") back to its old value. If we left "xp" unbalanced, we'd
have problems with diamond-shaped control flow:

    if (foo)
        printf("%d", 42);  /* decrements xp by 2 */
    localvar = 1;  /* we need to know localvar's offset relative
                      to xp, but now that depends on foo! */

Before this revision, David Given's algorithm was to assume that
every CALL should reset xp's offset to zero. But this isn't true!
Consider:

    printf("You scored %d out of a possible %d.\n", score(), 350);

This corresponds to an IC of

    PUSH 350
    CALL _score -> r0
    PUSH r0
    ASSIGN "%d %d" -> r0
    CALL _printf

Given's algorithm caused printf to "forget" any parameters that
followed a parameter containing a function call!

Fortunately, Barthelmann's IC foresaw this problem and included
a field "arg_list" which lets us see exactly which PUSHes correspond
to function arguments of the present CALL. In the above example,
the "arg_cnt" of CALL _score would be 0 and that of CALL _printf
would be 3 (only two of which would point to PUSH instructions).
So after CALL _score, we'd do no fixup at all, because its arg_list
didn't point to any PUSHes. Problem solved!

This new algorithm for adjusting the stack seems to work in all
cases; certainly it solves the problem in say_score() to which I
alluded earlier.