Fix the Z-machine implementation of the ran() function.
Apparently the Z-machine's @random instruction doesn't return a value
in the range [0,n); it returns a value in the range [1,n]. This was
essentially unnoticeable except when trying to hit a dwarf with the
axe; the dwarves were dodging with probability 2/3 instead of getting
hit with probability 2/3.

Also, random-walking dwarves would never take the first exit out of
any room, which meant that they'd tend to clump up together.

This is now fixed.