Fix the bear/axe interaction.
This was a bug in Knuth2002, fixed in Knuth2010. Throwing the axe at
the un-tame bear should immobilize the axe; but Knuth2002 got the
test backwards, so throwing the axe at the *tame* bear would immobilize
it (and then you'd have to go axeless the rest of the game).

Knuth2010 has fixed this, and I've adopted the fix.