Begin work on translating PLAT0550.
This file is Dave Platt's original Adventure source code; it's
written in a language Platt designed called "A-code". These days,
the major (only?) proponent of A-code is Mike Arnautov, who
maintains an A-code-to-C compiler called acdc:
http://www.mipmip.org/acode/

I plan to start with his conversion of PLAT0550, rip out all the
conditionally compiled stuff, and then polish the result. This
should be less prone to errors than my approach with KNUT0350,
where I started with a buggy version and did most of the conversion
by hand.