Fix an infinite loop if a FORCED move is blocked by a dwarf.
The program incorrectly allowed a dwarf to follow you from R_HMK to
R_SNAKED, and then block your (forced) motion from R_SNAKED back to
R_HMK. This led to an infinite loop of "You can't get by the snake!
A little dwarf with a big knife blocks your way. You can't get by
the snake!..." messages.

This was a bug in Knuth2002, fixed in Knuth2010. Oddly, the patch
in Knuth2010 exactly matches Woods' original code, down to the
comment; it's strange that the bug would have gotten introduced
in the first place.