Fix a bug with the caged bird in the endgame.
This was my error; Knuth2002 has it correct. The prop value of
the SNAKE may not matter, but the prop value of the bird is important,
because it may be picked up. Setting the bird's prop value to -1
meant that you could have an uncaged bird in your inventory; the
commands "GET CAGE. DROP CAGE" would result in a bird in the hand
and an empty cage on the ground. This is now fixed.

Also, I've updated my comment about reading the instructions.
Newbies get bonus lamp power, but they are docked 5 points; it's
impossible to achieve the rank of Adventure Grandmaster if you
start by reading the instructions.