Fix a typo-bug in the vbccz backend.
The vbccz backend incorrectly compiled this "hello world" program:

    #include <stdio.h>
    void hello(void) { puts("hello world"); }
    void bug(void) { puts("this should not be printed"); }
    void (*objs[])(void) = { bug, hello };
    void f(int i) { objs[1+i](); }
    int main() { f(0); return 0; }

The issue appears to have been a typo-bug in push_value().
We meant to check obj->val.vlong==0 "is this constant-expression
of the form &var as opposed to &var+42", but instead we were
checking obj->v->offset==0 "does var have a stack offset of zero",
which of course is nonsense for any EXTERN or STATIC variable.

After this fix, not only does the above program give correct results,
but a lot of code gets smaller; many load-address computations that
previously used two instructions
    @add _hash_table 0+0 -> sp;
    @add sp r0 -> r0;
are now using one instruction.
    @add _hash_table r0 -> r0;
This seems to cut the size of "advent.z8" by about 2 kilobytes!