Undo another of Knuth's quiet corrections to Woods' grammar.
The "little axe which missed, cursed, and ran away" is called out
in Dennis Jerz's review of Will Crowther's original code,
http://digitalhumanities.org/dhq/vol/001/2/000009/000009.html
The grammar must have grated on Knuth, or perhaps Woods even
agreed to the fix; but I prefer to stick as closely to the original
messages as possible, in this port. The odd wording is preserved
in Platt's version.