Continue work on ADV550-from-scratch.
The items are now mostly implemented, and the actions TAKE and DROP
(including TAKE INVENTORY), and the cave connections as far as Bedquilt.
I've added a z-machine/ directory and Makefile so that I will find out
quickly when advdat.c hits the limit of a .z8 file for code and/or data
(or the internal limits of vbccz, for that matter). It hasn't hit those
limits yet, which is moderately surprising.