Fix dwarf/goblin/sword/blob descriptions. Understand F_INVISIBLE.
Okay, so I wasn't quite done; I'd forgotten to go back and finish
those TODOs in describe_object().

I'd also introduced a bug by misunderstanding the purpose of
F_INVISIBLE; it should actually be named something more like
F_OMIT_PRECEDING_BLANK_LINE. You see, the code in look_around()
tries to put a blank line in between the room description and the
list of objects found there; but several objects have a description
of NULL, which means they don't print anything. It would be ugly to
have the extra blank line show up in those cases, so Platt tries to
give those objects F_INVISIBLE (meaning, "I'm not going to print
anything, so don't put a blank line in front of me").

But in the special case of DWARF, Platt wants there to always be
a blank line in front of the message "There is a threatening little
dwarf in the room with you!", even if it's the last object in the
room; and he *doesn't* want there to be *two* blank lines in front
of that message. So he gives the DWARF object the F_INVISIBLE flag,
and has each of the dwarf's actual descriptions begin with a newline.
This has the desired effect, but man is it confusing.

It turns out that by moving all the treasures into a block between
MIN_TREASURE and MAX_OBJ, I've actually broken the aforementioned
DWARF code. Looks like I'm going to need to add a valued() function
after all --- which of course I should have done in the first place,
since the goal is to mimic Platt's output exactly, including the
ordering of objects on the ground.