More bugfixes.
Added a return value for STAY_STILL, in order to distinguish "stay
still" from "moving to the same location" (such as in Witt's End).

Fixed the behavior of "AXE THROW, GET".

Fixed the initialization of some F_SCHIZOID objects; the dragon
wasn't properly showing up in R_SECRETCYNNE2, for example.