Fix some bugs related to Limbo and darkness.
The command "LOOK" called look_around(), forgetting to check whether
you were in darkness. This was sort of the reverse of my lamprey() bug,
where I factored out some code and then duplicated it; here I assumed
I had factored out the darkness-checking code when in fact I had not.

R_LIMBO needs a room description after all, because it is possible
to get there via the wizmode command "LPSD LIMBO". I actually want to
do this sometimes, e.g. to get the ring without having to do all the
puzzles in between. (But of course it's also possible to "LPSD FACES"
without getting the ring.) A few cases in describe_object() are also
reachable if you are a wizard in Limbo, so I've modified their
assertions.

In order to make "BACK" work from Limbo, I've moved F_NOBACK from
R_LIMBO to R_YLEM, and made oldloc = R_YLEM instead of R_LIMBO at the
start of the game or when resurrecting.

I'd cut-and-pasted the synonyms "volca" and "geyse" from ODWY0350;
those need to be full words in ODWY0550. And the "walking on lava"
message forgot to return STAY_STILL.