Fix a bunch of bugs related to fog and flags.
All of the upchuck() logic was getting bypassed because I forgot to
remove the check for the F_THROWER flag (which I decided doesn't
need to exist).

A few other room flags were set incorrectly. I think I've verified
all of them at this point, and the object flags too.

The visibility check in phog() was backwards; the glow showed up
exactly iff it wasn't supposed to, which made that puzzle a lot
easier than it was supposed to be! Also, the glow wasn't showing
up on the very first turn you moved into the fog-filled room; in
Platt's version it does. This is now fixed.

In the non-Z-machine C version, look_around() was segfaulting
anyway because R_PLAIN_2's descriptions were both NULL. And this
is assuming you could get past the assertion-failure in
describe_object() that couldn't handle objs[FOG].prop==8 (its
"default" invisible state). These are now both fixed.

Added a few missing blank lines involving Rover and the basilisk.