Extend the puts/printf optimization to handle "%d".
We're now optimizing printf("%d") into "@print_num". This applies even
when the "%d" specifier is embedded in a longer string. I've also mildly
refactored the loop so the base cases are a little bit clearer.

I've lowered the ceilings in the Makefiles to match the new "system
requirements" of Adventure and Adventure 550. The games' behavior
should remain unchanged.