Fix a couple of messages involving OPEN/FREE.
"OPEN" with no object, in the presence of the already-opened flask,
gave a bad message. Fixing this in a more robust way (by associating
each object with a default spelling, for example) is on my to-do
list.

The messages given by "DROP BIRD" were logical, but Platt's kernel
doesn't necessarily have logical behavior. Fixed "DROP BIRD" to
match the observed behavior of Platt's code.