Fix the description of the beads, and make 550 points achievable.
I'd somehow left out objs[BEADS].desc[1], so the beads would become
invisible when you dropped them for the first time. I didn't notice
this until the pirate stole them and I spent quite a while puzzling
over the pirate logic before realizing what had happened.

Premature optimization is the root of all evil. I had redefined
F_NOTINCAVE to include F_NODWARF, for ease of initializing the
location flags; but that was causing R_FACES to pass the "not in
cave" check in the endgame (since F_NODWARF & F_NOTINCAVE was
non-zero), which incorrectly reset its F_BEENHERE flag, which made
it impossible for anyone to achieve the full 550 points (because
you get 10 points for visiting the Valley of Stone Faces). I've
removed my clever definition of F_NOTINCAVE; problem solved.

After this revision, 550 points is achievable; I've done it
(although only by piping in a walkthrough, not by typing all the
way through on the keyboard, and *certainly* not without using
spoilers for all the mazes).