Implement a more foolproof default_to_something().
Specifically, each item now has a "default spelling" (whichever
spelling was the first one defined for that object).
The default_to_something() routine now strcpys that default
spelling into word2.text whenever it updates word2.meaning.
This ensures that word2.text is always meaningful, which
should help to avoid any further bugs like the one I fixed
in revision 17041cf8 with "OPEN [the flask]". This change also
lets me remove the hack I added in that revision to deal with
the flask as a special case.