Teach adv550 the "ignoreable" words from Platt's version.
This new code corresponds to line 164 in "PLAT0550/exec.f" and lines 33-34
in "PLAT0550/ADVENTURE.ACODE". Certain words, when they appear on the input
line, are simply ignored by the parser. This cleverly makes "GET IT" act the
same as "GET", for example. It's a very blunt instrument; you'll also find
that "THE A GET IT THEM LAMP" acts the same as "GET LAMP".