Fix some bugs uncovered by Clang warnings.
Whatever gave me the silly idea that MIST and TREES were objects?
Of course they're verbs!

I took this as an excuse to remove my examine_trees() code, which was
really an unwarranted bug-fix to Platt's code, and I'd been having
second thoughts about it for a long time.

I've also added "default: break;" clauses to the ends of a few
switch statements, which makes the compiler happier. But I haven't
done this consistently throughout, so -Wno-switch is still needed.