Undo my spelling corrections in this branch.
Rather than continue trying to split the difference between
faithfulness to the original versions and faithfulness to English
spelling and correct logic, I'm branching the project.
The "original-bugs" branch will contain the original games
exactly as before, but with the spelling errors restored (hence
this commit); the "master" branch will contain the games as I
think they're meant to be played, which will include em dashes
and a few bugfixes to Platt's code that I'd been putting off.