Decouple object-words from zero-based array indices.
ObjectWord enumerators are used for their numerical values
in two places: the objs[] array and the "cond" field of the
travel table. Subtract MIN_OBJ from the index in each case;
this will enable us to shift the ObjectWord enumerators up
so that we can fold all of the *Words into one big enumeration.