De-overlap the *Word enums; SOUTH no longer equals KEYS.
Most of the necessary changes were squashed into my previous
commit. However, one sneaky thing Knuth did was the assignment

    obj = *word2;

This looks nonsensical at first, but in fact it's punning on
the fact that *word2 is '\0' iff there are no more words on
the line, and 0 happens to be NOTHING, and we follow this line
with a test that only cares if obj is NOTHING or non-NOTHING;
the exact value of obj doesn't matter.

I've fixed this by inlining more of the logic, and will clean
it up a little more in my next commit too.