Fix the "FEED [dead] BEAR" bug.
KNUT0350 explicitly redirects "FEED BEAR" to "EAT BEAR" when the
bear is dead... but due to a sneaky switch-case fallthrough, the
message that ends up getting printed is default_msg[EAT], not
the normal "EAT BEAR" message. I had misread this when I ported
Knuth's code; hence my confused TODO comment here.

I finally checked this behavior against WOOD0350, and was surprised
to see that Woods' version of this code is perfectly straightforward,
with no sneaky control flow. If the bear is dead, print "Don't be
ridiculous!", full stop. I've fixed the mistake in my code.

(Notice that if the bear is dead, the food must already have been
consumed, and therefore will never be "here".)