Remove a TODO comment.
Knuth's code uselessly sets "k" to 0 or 1 here, which puzzled me;
but I see from WOOD0350 that Woods' code actually used "k" as an
index into a table of messages identical except for the word
"clam" (resp. "oyster"), and I guess Knuth was doing a literal
translation even though "k" is dead in KNUT0350; Knuth uses the
same "clam_oyster" local variable as I do, except that of course
he makes it a preprocessor macro instead of a real variable.

So, nothing needs to change, but I can now remove the comment.