Remove extra blank lines when climbing the plant.
The extra blank lines are caused by the double forced move
through R_CHECK. This appears to be a bug in Knuth's version:
he has R_CHECK's description as "", and prints the room
description unconditionally in Section 86. Contrariwise, in
Woods' original version, the description of room 31 is ">$<",
which causes it to be completely suppressed.