Refactor the BACK command.
In Knuth's code, the BACK loop would look for the first instruction with
a destination of (the-place-we're-trying-to-get-to), or, failing that,
it would take the *last* instruction with a destination that would
FORCED_MOVE us there. However, this complexity is inessential; I've
refactored it to simply take the first instruction that will get us to
our destination by any means.

The only place affected by the change was R_THRU/R_DUCK, the "low wide
passage parallel to and north of the Hall of Mists". If you go NORTH
from R_WFISS, you are forced through R_THRU and wind up at R_WMIST.
From that state, BACK must be equivalent to EAST, not NORTH. (Well,
it must be EAST if we're going to preserve Woods' original behavior,
anyway.) To make that happen, we have to swap a couple of make_ins()
statements in build_travel_table(). Voila, original behavior completely
preserved, and another "goto" eliminated!

Also, improve the comments on steal_all_your_treasure().