Remove the verb ABSTAIN, and set NOTHING to 0 instead of 200.
In Knuth's original port, ABSTAIN was the ActionWord with value 0
and NOTHING was the ObjectWord with value 0. Now that all "real"
words have values greater than 100, it doesn't make much sense to
have two different "no-input" words. So, I've put NOTHING back to 0
(making it WordClass_None) and removed ABSTAIN completely.

Notice that the conditional movement instructions (only_if_toting,
only_if_here, unless_prop) won't work correctly with an argument
of MIN_OBJ. That's not a problem as long as MIN_OBJ is KEYS, since
no conditional instruction in this game is triggered by the keys.