Fix the response to "FIND GULLY".
I'm somewhat disappointed that neither GCC nor Clang warned me
about this bug. FIND and INVENTORY are not MotionWords, but
ActionWords. Knuth's code gets this right; I introduced the bug
during my original port. The gist of Woods' code is that if the
player uses a motion word but *also* uses the word FIND (or
INVENTORY, for some reason), then we give the "I cannot tell you
where remote things are" message in preference to the "I don't
know how to apply that word here" message.

Due to what was probably sloppiness in Woods' original version,
the commands FIND CRAWL, FIND XYZZY, and FIND PLUGH do *not*
give the "I cannot tell you..." message when they fail.

Also, notice that FIND GULLY will act the same as GULLY by
itself, if the motion word GULLY is in fact an exit from the
current room. The new codepath is executed only if "mot" wasn't
found in the travel table for the current room.