Remove one Knuth-added fiddle with lost_treasures.
Compare to Crowther and Woods' line number 30310. If you collapse
the troll bridge without seeing the spices, Woods cuts you some
slack. Knuth added a check for the golden chain as well; but the
only way for that code to matter is if the player deliberately
entered the barren room in the dark, fed the bear in the dark,
and unlocked the chain in the dark (because that's the only
path by which you can get the bridge to collapse). So that check
is not only inauthentic but rather silly. I've removed it.