Woods' message was simply "You can't unlock the keys."
Knuth changed this message since it's given both for UNLOCK KEYS
and LOCK KEYS (and OPEN KEYS and CLOSE KEYS, since those verbs
are synonyms of LOCK and UNLOCK). The original game simply said
"You can't unlock the keys." regardless of which of those two
verbs you used to interact with them.