Factor out forbidden_to_pirate() as a function.
Knuth rearranges the rooms so that all the rooms forbidden to the
pirate happen to be greater than R_PIRATES_NEST. But that clever
pun isn't going to work for Long's version, so I might as well
factor it out now. As a bonus, the code is a little clearer.

No change in behavior.