Refactor and fix a bug with FILL VASE.
The behavior of FILL VASE (with water or oil) is broken in a different
way in every version of the game I've seen. In Woods' original, trying
to fill the vase while the pillow is on the ground beside you will
shatter the vase, then cause the vase to magically repair itself when
the game tries to DROP it next to you. In Knuth's port, the vase will
permanently shatter, but it remains in your inventory (as "Ming vase")
until you drop it. In McDonald's "Adventure 6", one of the authors clearly
noticed the bad behavior and tried to fix it, but you still end up with
a bogus message ("There is nothing here with which to fill the vase.")
in addition to a correctly shattered and dropped vase.

As part of this refactoring, I've removed the variable "suppress_ok_message"
from attempt_drop(), in an attempt to de-spaghettify things.