Fix more bugs in my map of the maze "all alike".
The connection from 53 to 51 should be west, not east;
there was a missing loop to the north of 42; and room 87
was missing entirely. Also, I've added little stalactite
icons on the three rooms (45, 50, and 53) into which you
are dropped at random when climbing down from the "top of
stalactite" room.

I've also verified that Platt's "Adventure 550" faithfully
reproduces this maze.