Fix a bird bug, and rename TROLL2 to NO_TROLL.
During my refactoring of attempt_drop(), I mistakenly removed the
uncaging of the bird when it's vaporized by the dragon. If we don't
uncage the bird properly, then dropping the cage will magically
put the caged bird back in the current room.

Also, I've reintroduced Woods' JUGGLE subroutine, which moves an
object back to the head of the object-list for its current location.
Knuth just uses "move()", which is equivalent, but I like the extra
self-documentation of using "juggle()" to indicate this cosmetic
shuffling of object descriptions.

I've also added a few sanity-checks in carry() and drop(), and
renamed Knuth's TROLL2 (Woods' "phony troll") to NO_TROLL. This
is the object whose description is "The troll is nowhere to be
seen."