Replace R_CRACK, R_CANT, R_SNAKED, and R_SEWER with remarks.
This has been on my to-do list for a while. After my refactoring
of the BACK code, I've convinced myself that indeed there is no
observable difference between these four rooms and the four remarks
I've replaced them with.

Perhaps R_UPNOUT could also be replaced with a remark, but I
believe that *would* have observable (if beneficial) effects,
via interaction with the BACK command.

Will Crowther's original source code didn't have any provision
for reusable "remarks"; that feature was added by Don Woods. When
Crowther wanted to rebuff the player with a message such as
    YOU DON'T FIT DOWN A TWO INCH HOLE!
he would create a new room with a forced move. (A forced-move
room, of course, can only have one exit.) Woods kept Crowther's
mechanism for forced moves, but also added "remarks" so that the
same rebuffing message could be used in multiple locations. For
example, Woods changed Crowther's message to
    YOU DON'T FIT THROUGH A TWO-INCH SLIT!
and made it a remark, so that he could use it at the "pit with
stream" location as well as aboveground. But for the few messages
that didn't specifically *need* to be reusable (CRACK, CANT,
SNAKED, and SEWER), Woods left them as pseudo-locations.

This change saves about 260 bytes of RAM on the Z-machine.