The motion-verb NOWHERE is impossible for the player to employ.
At least in Knuth's version, and mine, it's impossible for execution
to reach the travel-table codepath with mot==NOWHERE, because of
the several places that use NOWHERE as a special value meaning
"don't move" or "look around".

The mysterious NOWHERE connection appears in Crowther's original
source code. Personally, I suspect that "21" (NOWHERE) was a typo
for "23" (PASSAGE), which otherwise doesn't go anywhere from
the cobble crawl despite its being described as a "low PASSAGE"
with a "dim light at the east end of the PASSAGE". It would make
perfect sense for the verb PASSAGE to take you east.