Change the "vase broken" prop value from 2 to 1.
I had overlooked this unused object-description when I refactored
a lot of object-descriptions into regular printfs. There's no
reason to keep the message around as a description of the vase,
since it's printed only once (when you drop the vase on the pillow).
Then, in the spirit of my long-ago compression of the beanstalk
prop values, I've decided to compress the vase's prop values too.
The prop value for "worthless shards" is now 1, not 2.

I've also fixed a comment about DROP VASE. It turns out that I
was misreading Woods' code; when he compared LOC to PLAC(PILLOW)
and then checked AT(PILLOW), that *wasn't* redundant; what would
have been redundant is comparing LOC to PLACE(PILLOW). PLAC(PILLOW)
is a synonym for R_SOFT, so in fact Woods' code matches Knuth's
exactly.

Should be no change in behavior.