Remove a Knuth-added pun(?) from Witt's End.
Woods' message was
    YOU ARE AT WITT'S END.  PASSAGES LEAD OFF IN *ALL* DIRECTIONS.
Knuth changed it to
    You are at Witt's End.  Passages lead off in "all" directions.

The player might dismiss this punctuation as just another incorrect
use of "scare quotes" for "emphasis", but in fact I think Knuth may
have intended the quotes as a hint. Passages lead away from Witt's
End in all directions *except one*; hence not all directions, but
merely "all" directions. If so, this was a pretty clever little
modification, in the spirit of the original game.

However, in the interests of presenting a game as close to the
original as possible, I'm returning to Woods' wording.