Rearrange the motion vocabulary in original order.
Having the motion words in the same order as Woods' and Long's
will make it easier to showcase the differences between the two
versions, and also make it easier to verify that my list has
no unintentional differences from the original.

In fact my list *did* have an unintentional difference from
Knuth's; I was missing the synonym ASCEND for UP! This is now
fixed.