Rename CRYSTAL and BRIDGE to be more distinct.
Knuth's CRYSTAL is now FISSURE, and BRIDGE (the troll bridge) is
now CHASM. (CHASM isn't very distinctive, but I didn't like RICKETY
as an identifier.)

These identifiers match the names used by Woods, Long, and Platt,
and are already in use by my ODWY0550/. This also alleviates
ambiguity in "Adventure 6", which adds yet *another* chasm and
bridge (which is named CLAY) and a Crystal Palace (R_CRYSTAL)
unrelated to the crystal bridge.

No change in behavior.