Refactor is_treasure(t) and various objs(t).base references.
Woods cleverly puts all the treasures at the end of the list of
objects, so that "is this a treasure?" corresponds to "is this
object numbered greater than or equal to 50?". Knuth follows this
approach. But in "Adventure 6", we have too many new non-treasure
objects to keep following this approach, so Long introduced a new
object flag (bit 14) and function (TREASR) to distinguish
treasures from non-treasures. Let's switch to that approach.

One interesting thing about this change is that RUG_ is no longer
considered a "treasure", internally. This doesn't seem to have
any effect on gameplay. (I briefly wondered if looking at the
old rug from both sides would have given you 2+2 points instead
of just 2, but no, it doesn't.)

I've also cleaned up some other references to objs(t).base
while I'm at it. I don't think there's really any need to fix
the base of RUG_ before sending it off to R_LIMBO, and in
attempt_inventory() I don't think it's possible for the player
to be toting an immobile object.

Should be no change in behavior.