Rename Knuth's PONY to (vending) MACHINE.
Knuth explains in the literate comments that the identifier
PONY is a reference to the "Prancing Pony" at the Stanford
AI Lab (which was itself a reference to "Lord of the Rings").
However, this reference is original to Knuth; Woods' version
simply used the identifier VEND, and made no reference to
ponies, prancing or otherwise.

In the interest of not bewildering my readers, I've changed
PONY to MACHINE in all versions. (But I've left R_PONY alone,
because why not?)