Fix some minor grammar issues in the intro and hints.
Knuth quietly changed "...is useful in discovering..."
to "useful for". This may or may not have been intentional.

Also, he intentionally spells out "commands of one or two
words" (versus Crowther's "OF 1 OR 2 WORDS"), which changes
the linebreaks throughout the rest of that paragraph. I don't
think it's worth the gratuitous difference, so I'm going back
to "1 or 2 words".

I've also changed Knuth's word "FORTRAN" to mixed case, which
is easier on the eyes (in case anyone actually reads all the
way to the end of the credits).