Initial commit of ODWY0501/.
This is a work in progress. I started with ODWY0350; replaced
the travel table, objects, and vocabulary; replaced the listen()
routine with getwds() (which is currently instrumented with
debugging printfs); and ripped out almost all of the ODWY0350
game logic, since I know a good fraction of it will need to be
reimplemented anyway.

The program compiles and runs, and you can move around following
the travel table, but no verbs are implemented so you can't
GET anything or even LOOK. (But you can QUIT!)