Revert some of Knuth's grammatical changes to messages.
The "low rumbling" change just seems gratuitous. But I've also
added a block comment explaining that I agree with some changes
Knuth made to the SW end of the repository; Woods' version has
some bad linebreaks and repetition of the word "which". For
the original messages, see the original-bugs branch.