Continue working on ODWY0501: CLOSE and POUR.
Implemented the verbs CLOSE and POUR.

Tracked down the exact cause of the "cask full of cave drawings"
bug I'd observed in Long's version, explained it in comments, and
patched it.

Implemented Easter eggs for strange places the sapphire could be.
(This breaks from my "originalist" goal, but after patching so
many bugs in Long's original code, I figure I'm making more of
a "reimplementation" than a "port" at this point, and I deserve
to have a little fun here and there.)