Continue working on ODWY0501: verb EAT.
Implemented the verb EAT. Investigated several amusing bugs
related to the Alice puzzle; for example, it's possible to strand
yourself in R_CRAMPED without any inventory or items in the room
with you, by vanishing the key into the chest-in-bottle-in-chest
and then eating the cakes, or of course by vanishing the cakes
to begin with.

Implemented forbidden_to_pirate(), and cleaned up some movement
and pseudo-movement stuff. Dwarves and pirate are still largely
unimplemented.

Investigated the behavior of the word "LIGHT" as a verb and noun.
I believe I know what the program is doing, but I haven't figured
out how it corresponds to the source code!