Better solution for indentation; verb LIGHT; getwds() fixes.
I've ripped out the indent_if_needed() routine and replaced it
with something closer to how I hope this will work on the
Z-machine (with compiler support, that is). I've now got a
global function set_indentation(x) that indents all output
by x spaces; this is done by re-#defining "puts" to a routine
of my own. I've made sure that "printf" never needs to be
redefined in this way, because THAT would get ugly!

My TODO comment about a possible off-by-one bug in getwds()
was right on the money. "GET LAMP AND SACK" would trigger
the response "What?" instead of properly dealing with the
two-object command. This is now fixed.