Fix some original bugs in DROP.
DROP ROD would not correctly disambiguate if you were toting the
rusty-marked rod inside a container. Not that that's possible
(since there are no general-purpose containers in the endgame);
but might as well be consistent.

DROP WATER would behave incorrectly if both bottle and cask were
full of water, but only the cask were toted. (It would redirect
to DROP BOTTLE and give the response "You aren't carrying it!")

Also use move(obj, loc) instead of drop(obj, loc); the latter
assert-fails if obj is toted inside a container.