Continue working on ODWY0501: SWEEP, BLAST, clocks.
Long's game adds a third clock to Woods' two. The third
clock counts down a random number of turns (between 20 and 40)
when you're in the repository; at the end of that timer the
phone begins to ring, and seven turns later the dwarves are
awakened by the noise --- game over.

I've removed the wizmode commands WIZ, MAP, GATE, and PIRLOC;
these are implemented in MCDO0551 but WIZ's implementation
is commented out so that there is no way to get into wizmode.
This means that all MAP, GATE, and PIRLOC will do is print
"OK" (that is, what they do outside of wizmode).