Continue working on ODWY0501: verbs EXTINGUISH and WAVE.
Also some more Wumpus stuff and cave-closing stuff.

I've noticed that Long's "close the cave" logic is identical
to Woods', except that Long also closes the elfin door. (He
forgets to lock it, but I've fixed that.) Long doesn't kill
the sleeping dog or the chattering gnome. But then, Woods
didn't kill the dragon either, so I guess that's okay.