Continue working on ODWY0501: BLOW, PLAY, DIAL, CALL, bugfixes.
I've realized that Long's room R_ESHORT is completely unnecessary.
It is inaccessible, both by the player and by the dwarves, except
that the player can get there by exploiting the cask bug:
PUT CROWN IN CASK gives you the ruby slippers prematurely, which
enables you to enter the Gothic Cathedral section via the Rainbow
Room. Since I'm not emulating the cask bug, there's no way to see
R_ESHORT, and therefore I've removed it from my port of the game
(to save space).

The ECHO object therefore doesn't need to be schizoid; I've
removed the ECHO_ object.

I'd mistakenly added a DUSTY_ROCKS object in my first pass; really,
the dusty ROCKS are just a synonym for the CARVING, and don't have
a separate noun in Long's game.

Long fiddles with PROP(POLE) in several places, but I think this
is unnecessary. PROP(BOAT) is the important variable as far as
the travel table is concerned.

I've factored out wake_the_wumpus() and remobilize_axe().  I use
the latter function consistently, even though Long doesn't always
check the axe's location before remobilizing it. (For example, if
you throw the axe at the bear and then calm the dog, the axe will
be mobile the next time you visit the barren room.) I consider
this a bugfix.

I've decided to prohibit putting the pole inside the chest, so
as to get rid of that loophole in Long's general "can't be in
the boat without the pole" invariant. This is a bugfix similar
in justification to my prohibition on putting the bear inside
the chest (albeit not as much fun to exploit).